+++ tests/input_files/series/bar/06.js
@@ -100,8 +100,8 @@
         symbol: 'circle',
         width: 48
       },
-      x: 'some category',
-      y: 123
+      y: 123,
+      name: 'some category'
     },
     {
       direction: 45,
@@ -203,8 +203,8 @@
         symbol: 'circle',
         width: 48
       },
-      x: 'some category',
-      y: 123
+      y: 123,
+      name: 'some category'
     }
   ],
   id: 'some-id-goes-here',