+++ tests/input_files/series/boxplot/01.js
@@ -107,8 +107,8 @@
         symbol: 'circle',
         width: 48
       },
-      x: 'some category',
-      y: 123
+      y: 123,
+      name: 'some category'
     }
   ],
   id: 'some-id-goes-here',