+++ tests/input_files/series/data/range/01.js
@@ -97,5 +97,5 @@
     symbol: 'circle',
     width: 48
   },
-  x: 'some category'
+  name: 'some category'
 }