+++ tests/input_files/series/heatmap/02.js
@@ -100,8 +100,8 @@
         symbol: 'circle',
         width: 48
       },
-      x: 'some category',
-      y: 123
+      y: 123,
+      name: 'some category'
     }
   ],