+++ tests/options/series/test_pictorial.py
@@ -112,8 +112,8 @@ STANDARD_PARAMS = [
               'symbol': 'circle',
               'width': 48
             },
-            'x': 'some category',
-            'y': 123
+            'y': 123,
+            'name': 'some category',
           },
           {
             'dataLabels': {
@@ -420,8 +420,8 @@ STANDARD_PARAMS = [
               'symbol': 'circle',
               'width': 48
             },
-            'x': 'some category',
-            'y': 123
+            'y': 123,
+            'name': 'some category'
           },
           {
             'dataLabels': {