+++ CHANGES.rst
@@ -1,3 +1,16 @@
+Release 1.3.0
+=========================================
+
+* **ENHANCEMENT:** Modified the way that data points are serialized to JavaScript literal objects. Now, they are serialized to a JavaScript array if their configured properties are those that Highcharts (JS) supports in JavaScript array notation. Otherwise, the code falls back to serialize the data point as a JavaScript object literal. This change is intended to improve performance and reduce the size of the serialized data. (#77)
+* **ENHANCEMENT:** Added ``__repr__()`` method for Highcharts Core for Python classes (#76).
+* **ENHANCEMENT:** Added ``__str__()`` method with special handling for difficult-to-read classes (#76).
+* **ENHANCEMENT:** Added ``Chart.get_script_tags()`` to retrieve Javascript ``<script>`` tags (#78).
+* **ENHANCEMENT:** Added ``utility_functions.to_snake_case()`` function.
+* **BUGFIX:** Fixed incorrect serialization of datetime and Pandas ``Timestamp`` objects in ``.to_dict()`` and ``.to_json()`` (#74).
+* **BUGFIX:** Fixed incorrect serialization of ``EnforcedNull`` in ``.to_dict()`` and ``.to_json()`` (#75).
+
+------------------
+
 Release 1.2.6
 =========================================